dt-bindings: pinctrl: qcom: msm8953: make function property conditional
Some pins on MSM8953 SoC have only one function. Which means that there
isn't going to be any value to set "function" property to. Only other
settings can be modified for those pins.

This patch adds condition to disallow "function" for sdc and qdsd pins
and make it required only for the rest.

Signed-off-by: Vladimir Lypak <vladimir.lypak@gmail.com>