drm/msm/a5xx: properly clean preempt record after suspend
Some fields of preempt_record aren't zeroed on resume: "data" and
"info". This is the cause of faults which happen when we try to switch to
the ring that was active last before suspend. Either way we better
clear both of those fields on all rings like it's done in downstream
driver (KGSL).

Signed-off-by: Vladimir Lypak <vladimir.lypak@gmail.com>